Fix breadcrumbs
We always show a "Files" breadcrumb as the root node.
The API returns an empty array for parents for the root node, and
subdirectories have the root node in the parents list. This makes sense
from the APIs perspective. But here in this for-loop we've already
accounted for the root node, so we need to filter it out.

Also the root node now has an ID, where previously it didn't. So we need
to hide the leaf node if it is the root node since again we've already
accounted for it.