Fixed project & internship list flipping back to "my list"
This happened when the page was loaded with query param all=true already set.
`queryParamChanges` was called before `extraInputs` was subscribed too.
This means that the call to `listChanges.next` with the correct value was ignored
because there was no listener (subscription for `listChanges`.
Changing `listChanges` to a BehaviorSubject ensures that the correct value is always
available where there is a listener or not.

This code could probably be simplified more, but meh. It's going away soon.