Refactor \`FileActionsContext\` to allow passing in list of actions to display
The previous iteration of this required a \`context\` prop to be passed in to
the Provider, and a `canEdit` prop to be passed in to `FileActionsMenu`. These
values were used to determine which actions to show in the File Actions Menu,
since Engagement Files have different requirements than Project files. But this
felt really brittle, so we refactor to give control of which actions are shown
to the consumer of `FileActionsMenu`.

The tricky part of this has always been that one of the consumers of
`FileActionsMenu` is `FileVersionItem`, which is created by `FileVersions`,
which is created by `FileActionsContext`, which is controlled through
`FileActionsMenu`. So this updates the `handleFileActionClick` function of the
`FileActionsContext` to allow `FileActionsMenu`, when it opens `FileVersion`,
to pass along its list of permitted actions.

We also update the typings in `FileActionsContext` to accept the types of the
Engagement File fragments.