De-duplicate uploading code between `useUploadProjectFiles` and `useUploadEngagementFiles`
These two hooks both essentially had the same "useUploadFiles" and typings, and
differed only in the handling of the completed upload. So we break out a new
`components/files/hooks/useUploadFiles`, where we centralize the typings and
just add a new `handleUploadCompleted` property in the `input` object passed to
the returned function. Then we update all consumers of the two original hooks
to consume the new hook and just pass the correct post-upload handler.