Remove `Grid` around detail cards; Supply loading state to Budget card
We were trying to fix a problem where the Budget card's height did not match
the Files card's height during loading. We had originally fixed this by
wrapping all the cards in a `material-ui` `<Grid>` component, but that had the
side effect of making the top row of cards the same height as the bottom row,
which was undesirable.

The real problem was that the Budget card lacked a `loading` prop. Supplying
this fixed the original issue, and removing the `<Grid>` fixed the new issue.

While we were in here messing around, we updated the props for
`FilesOverviewCard` and `BudgetOvervieCard` to extend those of
`FieldOverviewCard`, which they are wrapping.