Remove select variant for now
It's proving to be difficult due to the fact it uses
"reflection" (component's props) on its children to
identify currently selected option & clones the children
to add more HTML props.

Adding symbol here as an option as a workaround for eslint.
I'm generally against this, but it seems like the best option here.
This allows the component to not have to remove the last
condition on variant to appease useless-condition rule.
I want to keep specific variant code under that condition.
Partly because the variants could be changed/added in future
and I don't want this other variant code to have to change just
because it's not the last condition now.