Configure "single port" for dev server
In reality, in dev, we are hosting the app on the defined port+1
and the webpack dev server is on the defined port and proxying
to the incremented port that our server is listening on.

This allows the browser to only use one port which prevents confusion
in errors occurring across origins.