Update list pagination merging to be more robust
Previously we used a sparse array so each page could be loaded out of order,
while maintaining sorting. However, this assumed a unchanging list,
which is not the case due to client side additions as well as server side
additions/removals/changes causing order to change.

Now we ignore which page we are loading and just merge under the assumption
we can sort the list ourselves based on the sorting config.
If we have sorting config then we can splice the new page at the appropriate
spot in the list. This has the benefit of removing deleted items if they were
in the middle of the new page.
If we don't have sorting config then we just append the new page.
We also unique the merged list so that unstable pagination or shifting pages
doesn't result in duplicate items in old & new spots.

There may be some unnecessary steps here, but hopefully it's resilient enough
to go to prod and we can iterate later when we have more info.