Upgrade & patch MUI date pickers to fix bugs
I think this fixes the glaring issues with the date field functionality.
Obviously it's not ideal to patch. Hopefully we'll be able to jump on
MUI v5 as soon as it goes stable and this will get better there.
This patch will hopefully bridge the gap until we get there.