Expand query string enum handling
Enum(List)Param now accepts a list of valid options,
any values not found in the valid list are ignored.

Values are automatically encoded to lowercase, but
this can be overridden with the mappingOverride arg.

This fixes handling of ProjectStatus.DidNotDevelop,
since previously we were just naively upperFirst-ing
the value (i.e. "Didnotdevelop").