Add type policy for our lookup by ID queries
Tell Apollo that these query operations return the type given with the ID referenced.
This allows a network operation to be saved if we've already cached the object.

Note that Apollo also checks that all fields being asked for exist as well,
so this only helps if the complete object needed was fetched another way.