Split gql schema generation from operations generation
This means that the schema files will be updated regardless of whether the operations will compile.
When a breaking change happens, we can now run gql gen to get the new schema.
Then we can see exactly what's wrong & needs to be updated with our operations.
After those changes are made we can run gql gen again to get the operations generated.