Prevent Dialog from closing when clicking away from DatePicker
It seems like there should be a better way to do this.
Keyboard navigation is already has this scope enforcement,
you'd think there'd be a builtin way to enforce scope with mouse.

While this is a bit of a workaround, I'm happy with implementation;
it doesn't feel hacky.
Though allowing any dialog children to prevent dialog closing could
be a double edge sword and should be handled responsibly.