Ensure cache updating for pinned list doesn't assume args
Technically these args could be anything, not the type specified.
It's all based on cached queries.
Additionally, Apollo seems to be dropping args from cache keys if they are empty objects.
So {input: {}} is simplified to {}
So when we JSON parse this we obviously don't have the object at input, even if the operation
requires it & we have a TS type for it.