Don't require the API PR to be successful when pulling the API GQL schema
The API workflow has the "gql breaking changes" job within the CI workflow we reference here.
Ideally that would be split out.
But we also don't really care for this logic. As long as the API can generate the schema,
we'll use it. Let the API PR worry about everything else.