Fix ProgressReportList click handler
A previous iteration (that I squashed) had the object at report,
and that's how it was before with material-table.
I went back and flattened it so the row shape was just the actual report object,
as this is how DataGrid intended it to be.

This was missed though because the row generic is not passed through to
DataGrid's props including this event handler prop. Hopefully that will be resolved someday by them.