Use current user's assignable roles when creating new users (#1383)
This fixes the UX where roles cannot be assigned while creating a new user.

We assume the assignable roles of the current user as the same as new users. I think this is safe although it's not the explicit API contract. The API can still do validation if there's an error here.

I also changed the list to just filter out un-assignable vs just disabling them. This seems like a better UX as the assignable could be short, and our roles are getting longer.