When merging new items to all-pages cache, limit item fields to only the key args
The cache stores the list as a list of refs, and that is really what we want to update, not the fields of individual items.
I'd rather use the lower-level `cache.modify()` API to convey this ref list change more directly.
But we have to use `cache.write()` since there will be multiple cache entries for each combination of filters/sorts/pages,
and we need to pick the "all-pages" one (based on `variables` / lack of `input` qualifiers).
Only write() can create a new entry as well, which is needed for this logic here.

Since we can't use `refs` with this API, we use limit the fields to the "keys args", which is the source of how the ref is created.
i.e. __typename & id become `User:1234`.

Why do this at all?
Apollo gives us field values here (both prev & next) that have been passed through the `read` type policies.
This means our temporal scalars have been converted from ISO to temporal objects.
Apollo has no way of knowing how to reverse that transformation
and thus stores these values with a different type after this logic completes.
By limiting to only key-args we assume that these fields will not be transformed and therefore safe to write back to the cache.