Fix DataGrid resetting to first page on last page
This happens I believe because we unset `rowCount` (due to MUI's own warning),
when flipping from server pagination to client pagination.
An effect is triggered to set the rowCount to the `rows.length`.
And maybe during this somewhere DataGrid thinks it needs fix the pagination model,
because the current state is invalid for a brief moment, resetting to page 0.

Instead, we'll just always handle the pagination ourselves.