Customize boolean filter handling to actual booleans instead of string representations
This is needed so to send the correct value types to the API.
We also changed the rendering to match enum columns.

It is a bit concerning that MUI has falsy checks for `FilterItem.value`
I believe we've replaced the ones that matter.
However, it's possible we'd want to revert this back to a string rep which is truthy values.
We could add another getter function to the column def to convert filter items to API values,
to also achieve this bug fix.

Co-authored-by: Rob Donigian <rob_donigian@tsco.org>