Stop implicitly parsing ISO date strings (#1660)
Now do it explicitly when/where needed.

It is too hard to get right currently.
The outstanding apollo client issue doesn't have any appearance of resolving this read issue.

This change better reflects reality, as we're no longer assuming luxon instances when the values are actually ISO strings.

Though I did cheat a little bit and say that they could be either an ISO string or a luxon instance, which isn't really true. But it seemed easiest to go with this for now, and then narrow to luxon instance explicitly with a function call.