Fix zone director form value
The form field path name needs to match the GQL input path name.
This is what the API uses when there are certain field errors.
This lets us transparently show API errors on certain fields.

Like for example in this case, we can show that the user
does not have the required director role.