Switch to `graphql-sse` library
This library uses `fetch()` instead of raw `EventSource`.

This has 2 things I want:

1. It does its own retry logic before communicating to Apollo
the stream is dead. This is seemingly redundant with our
retry link, however, the retry attempts are _reset_ on a
successful connection/emission. This allows streams to disconnect
and reconnect (within reason) infinite types transparent to the
client. This could happen due to network problems or due to
server restarts/deployments.

2. Custom headers are allowed, which are required for our
impersonation logic. (It might be nice to upgrade the API
to move/allow these in the extensions' object as well.)

3. This uses POST and should preflight / avoid CSRF blocks.

This does have to do a little work to parse the text stream
into events, but this doesn't mean it is slower that native browser.
It is also not like events are spammed either, this will be
a very slow trickle of data if anything.